- Adding gasPrice optional for more flexibility - Consolidate Types into 1 package - Adding instance var `title` back into provider
