--- packages/Arbiter/src/types.ts
@@ -1,28 +1,0 @@
-import {address,BNType, defaultTx, Filter,NumType} from "@zapjs/types"
-
-export interface SubscriptionInit extends defaultTx{
-    provider:address,
-    endpoint :string,
-    endpoint_params: Array<string>,
-    blocks : NumType,
-    pubkey:NumType,
-    from: address
-}
-
-export interface SubscriptionEnd extends defaultTx{
-    provider?:address,
-    subscriber ?:address,
-    endpoint : string
-}
-
-export interface SubscriptionType {
-    provider:address,
-    subscriber:address,
-    endpoint:string
-}
-
-export interface SubscriptionParams  extends defaultTx{
-    receiver:address,
-    endpoint:string,
-    params: Array<string>
-}
