+++ packages/Bondage/src/index.ts
@@ -1,6 +1,6 @@
 import {BaseContract} from "@zapjs/basecontract";
-import {BondageArgs, BondArgs, UnbondArgs, DelegateBondArgs, BondFilter} from "./types";
-import {Filter,txid,NetworkProviderOptions,BNType,NULL_ADDRESS,NumType} from "@zapjs/types"
+import {BondageArgs, BondArgs, UnbondArgs, DelegateBondArgs, BondFilter,
+    Filter,txid,NetworkProviderOptions,BNType,NULL_ADDRESS,NumType} from "@zapjs/types"
 const {utf8ToHex,isHex,toHex} = require("web3-utils");
 const assert = require("assert");
 const DEFAULT_GAS = 300000
@@ -272,5 +272,3 @@ export class ZapBondage extends BaseContract {
     }
 
 }
-
-export * from "./types";
